fix: Adjust reset password page.

+++ milestone-tracker-backend/supabase/migrations/20230515141104_add_proposals.sql
@@ -8,7 +8,7 @@ CREATE TABLE IF NOT EXISTS public.proposals
     challenge_id bigint,
     budget bigint,
     milestones_qty bigint DEFAULT 5,
-    funds_distributed bigint DEFAULT 0,
+    funds_distributed float DEFAULT 0,
     CONSTRAINT proposals_pkey PRIMARY KEY (id),
     CONSTRAINT proposals_challenge_id_fkey FOREIGN KEY (challenge_id)
         REFERENCES public.challenges (id) MATCH SIMPLE
